// src/types/globalTypes.ts
export interface Istyle {
  textAlign?: string;
  width?: string;
  maxWidth?: string;
  minWidth?: string;
  height?: string;
  fontSize?: string;
  padding?: string;
  paddingLeft?: string
  borderRadius?: string;
  borderBottom?: string;
  borderTop?: string
  borderLeft?: string
  borderRight?: string
  backgroundColor?: string;
  color?: string;
  border?: string;
}